WID-174. Refactor: divide ts_widget.py file into multiple modules
