+++ tvbwidgets/api.py
@@ -8,7 +8,9 @@
 from .ui.phase_plane_widget import PhasePlaneWidget
 from .ui.storage_widget import StorageWidget
 from .ui.head_widget import HeadBrowser, HeadWidget, HeadWidgetConfig
-from .ui.ts_widget import TimeSeriesWidget, TimeSeriesBrowser, TimeSeriesWidgetPlotly
+from .ui.ts.mne_ts_widget import TimeSeriesWidgetMNE
+from .ui.ts.plotly_ts_widget import TimeSeriesWidgetPlotly
+from .ui.ts.ts_widget_browser import TimeSeriesBrowser
 from .ui.pse_widget import PSEWidget
 from .ui.pse_launcher_widget import PSELauncher
 from tvbwidgets.core.hpc.config import HPCConfig
@@ -26,7 +28,7 @@ def plot_timeseries(data, sample_freq=None, ch_idx=None, backend='matplotlib'):
     """
 
     if backend == 'matplotlib':
-        tsw = TimeSeriesWidget()
+        tsw = TimeSeriesWidgetMNE()
     elif backend == 'plotly':
         tsw = TimeSeriesWidgetPlotly()
     else:
